refacor : Operation 클래스를 Calculator와 통합 후 제거
